debug asserve

+++ Projet_Crabot.X/lib_asserv/private/asserv.h
@@ -45,7 +45,7 @@ typedef struct {
     MotionState *state;
     MotionConstraint *constraint;
     int done;
-    int courbure;
+    float courbure;
 } SpeedAsserv;
 
 // Un asservissement en angle
